Extract page tags panel
